feat: add proficiency base module
